ares_init_options with ARES_OPT_UDP_PORT wrong byte order
Regression from c-ares 1.19.1, ARES_OPT_UDP_PORT and ARES_OPT_TCP_PORT are
specified from the user in host-byte order, but there was a regression that
caused it to be read as if it was network byte order.

Fixes Bug: #640
Reported By: @Flow86
Fix By: Brad House (@bradh352)

+++ test/ares-test-main.cc
@@ -27,7 +27,7 @@ int main(int argc, char* argv[]) {
       ares::test::verbose = true;
     } else if ((strcmp(argv[ii], "-p") == 0) && (ii + 1 < argc)) {
       ii++;
-      ares::test::mock_port = atoi(argv[ii]);
+      ares::test::mock_port = (unsigned short)atoi(argv[ii]);
     } else if (strcmp(argv[ii], "-4") == 0) {
       ares::test::families = ares::test::ipv4_family;
       ares::test::families_modes = ares::test::ipv4_family_both_modes;
